+++ sample_app/lib/pages/thread_page.dart
@@ -55,7 +55,6 @@ class _ThreadPageState extends State<ThreadPage> {
               parentMessage: widget.parent,
               initialScrollIndex: widget.initialScrollIndex,
               initialAlignment: widget.initialAlignment,
-              markReadWhenAtTheBottom: true,
               //onMessageSwiped: _reply,
               messageFilter: defaultFilter,
               showScrollToBottom: false,